fix: update, organize, document template rules.
- updated to latest sprig v3 libraries for template functions
- update to go 1.22
- centralized template creation on Engine for consistent functions & options.
- separted log.yaml rules from k8s.yaml
- documentation for template functions - package doc
- log but don't return traversal errors.

Korrel8r should return whatever it can find even if some rules fail to match or generate bad queries.
