+++ .bingo/golangci-lint.mod
@@ -2,4 +2,4 @@ module _ // Auto generated by https://github.com/bwplotka/bingo. DO NOT EDIT
 
 go 1.21.7
 
-require github.com/golangci/golangci-lint v1.57.2 // cmd/golangci-lint
+require github.com/golangci/golangci-lint v1.59.1 // cmd/golangci-lint
