+++ .bingo/variables.env
@@ -10,7 +10,7 @@ fi
 
 CONTROLLER_GEN="${GOBIN}/controller-gen-v0.14.0"
 
-GOLANGCI_LINT="${GOBIN}/golangci-lint-v1.57.2"
+GOLANGCI_LINT="${GOBIN}/golangci-lint-v1.59.1"
 
 KORREL8RCLI="${GOBIN}/korrel8rcli-v0.0.2"
