+++ .github/workflows/build.yml
@@ -14,15 +14,17 @@ jobs:
       - uses: actions/checkout@v4
       - uses: actions/setup-go@v4
         with:
-          go-version: "1.21"
-      - uses: actions/cache@v3
+          go-version: "1.22"
+      - uses: actions/cache@v4
         with:
-          path: ~/go
+          path: |
+            ~/go
+            ~/.cache/golangci-lint
           key: cache-${{ runner.os }}-${{ runner.arch }}
 
       - name: Lint
         run: |
-          make lint
+          make GOLANGCI_LINT_CACHE=~/.cache/golangci-lint lint
           git diff --exit-code
 
       - name: Test
