+++ pkg/graph/traverse.go
@@ -12,7 +12,7 @@ import (
 
 // Traverse rules on paths from start to goal.
 // Returns the subset of the graph that was traversed.
-func (g *Graph) Traverse(start korrel8r.Class, goals []korrel8r.Class, f func(*Line) bool) *Graph {
+func (g *Graph) Traverse(start korrel8r.Class, goals []korrel8r.Class, f func(*Line) bool) (*Graph, error) {
 	sub := g.Data.EmptyGraph()
 	bf := traverse.BreadthFirst{
 		Traverse: func(edge graph.Edge) bool {
@@ -26,12 +26,12 @@ func (g *Graph) Traverse(start korrel8r.Class, goals []korrel8r.Class, f func(*L
 		}}
 	bf.
 		Walk(g, g.NodeFor(start), nil)
-	return sub
+	return sub, nil
 }
 
 // Neighbours traverses a breadth-first neighbourhood of start.
 // Returns the subset of the graph that was traversed.
-func (g *Graph) Neighbours(start korrel8r.Class, depth int, f func(*Line) bool) *Graph {
+func (g *Graph) Neighbours(start korrel8r.Class, depth int, f func(*Line) bool) (*Graph, error) {
 	sub := g.Data.EmptyGraph()
 	atDepth := 0
 	current := unique.Set[int64]{} // Nodes at the current depth or above.
@@ -59,7 +59,7 @@ func (g *Graph) Neighbours(start korrel8r.Class, depth int, f func(*Line) bool) 
 		}
 		return d > depth
 	})
-	return sub
+	return sub, nil
 }
 
 // traverseEdge calls f(l) for each line l in the edge.
